+++ tests/IntegrationTests.js
@@ -28,7 +28,7 @@ gpii.loadTestingSupport();
 
 require("./DialogManagerTestDefs.js");
 require("./IntegrationTestDefs.js");
-require("./QssTestDefs.js");
+require("./qss/QssTestDefs.js");
 require("./SequentialDialogsTestDefs.js");
 require("./SettingsBrokerTestDefs.js");
 require("./StorageTestDefs.js");
@@ -227,11 +227,11 @@ if (gpii.tests.app.isInstrumented) {
 
 
 gpii.tests.app.bootstrapServer([
-    fluid.copy(gpii.tests.app.testDefs),
-    fluid.copy(gpii.tests.dev.testDefs),
-    fluid.copy(gpii.tests.timer.testDefs),
-    fluid.copy(gpii.tests.dialogManager.testDefs),
-    fluid.copy(gpii.tests.qss.testDefs),
+    // fluid.copy(gpii.tests.app.testDefs),
+    // fluid.copy(gpii.tests.dev.testDefs),
+    // fluid.copy(gpii.tests.timer.testDefs),
+    // fluid.copy(gpii.tests.dialogManager.testDefs),
+    fluid.copy(gpii.tests.qss.testDefs)/*,
     fluid.copy(gpii.tests.sequentialDialogs.testDefs),
     //fluid.copy(gpii.tests.shortcutsManager.testDefs), // NOT OK
     fluid.copy(gpii.tests.settingsBroker.testDefs),
@@ -241,5 +241,5 @@ gpii.tests.app.bootstrapServer([
     fluid.copy(gpii.tests.storage.testDefs),
     fluid.copy(gpii.tests.userErrorsHandler.testDefs),
     // fluid.copy(gpii.tests.gpiiConnector.testDefs),  // should be changed to match the new specification
-    fluid.copy(gpii.tests.webview.testDefs)
+    fluid.copy(gpii.tests.webview.testDefs)*/
 ]);
