+++ tests/fixtures/qssSettings.json
@@ -1,5 +1,6 @@
 [
     {
+        "id": "language",
         "path": "http://registry\\.gpii\\.net/common/language",
         "schema": {
             "type": "string",
@@ -34,6 +35,7 @@
         "tooltip": "<p>Change Windows language setting.</p><p>Works for menus, etc. in Windows, Edge, and many other applications that have Language Packs installed.</p>",
         "restartWarning": true
     }, {
+        "id": "screen-zoom",
         "path": "http://registry\\.gpii\\.net/common/DPIScale",
         "schema": {
             "type": "number",
@@ -54,6 +56,7 @@
         },
         "learnMoreLink": "https://morphic.world/help/qsshelp#screenzoom"
     }, {
+        "id": "text-zoom",
         "path": "appTextZoom",
         "schema": {
             "type": "number",
@@ -73,6 +76,7 @@
         },
         "learnMoreLink": "https://morphic.world/help/qsshelp#textzoom"
     }, {
+        "id": "high-contrast",
         "path": "http://registry\\.gpii\\.net/common/highContrastTheme",
         "schema": {
             "type": "string",
@@ -147,6 +151,7 @@
         "tooltip": "<p>Change the contrast between the text and background color, so text is easier to read.</p><p>Contrast changes will affect Windows, Edge, Chrome, and some other applications.</p>",
         "learnMoreLink": "https://morphic.world/help/qsshelp#contrast"
     }, {
+        "id": "read-aloud",
         "path": "http://registry\\.gpii\\.net/common/selfVoicing/enabled",
         "schema": {
             "title": "Read Aloud",
@@ -162,6 +167,7 @@
         "learnMoreLink": "https://morphic.world/basics/learn-more-about-quickstrip#readaloud",
         "tooltip": "<p>Turn “Select to Read Aloud” extension in Chrome browser on/off.</p>"
     }, {
+        "id": "screen-capture",
         "path": "screenCapture",
         "schema": {
             "type": "screenCapture",
@@ -178,6 +184,7 @@
         "learnMoreLink": "https://morphic.world/help/qsshelp#screen-capture",
         "value": ""
     }, {
+        "id": "office-simplification",
         "path": "officeSimplification",
         "schema": {
             "type": "office",
@@ -192,7 +199,8 @@
         "messageKey": "officeSimplification",
         "learnMoreLink": "https://morphic.world/help/qsshelp#office-simplification",
         "value": ""
-    },  {
+    }, {
+        "id": "volume",
         "path": "http://registry\\.gpii\\.net/common/volume",
         "schema": {
             "type": "volume",
@@ -210,6 +218,7 @@
         "learnMoreLink": "https://morphic.world/help/qsshelp#volume"
 
     }, {
+        "id": "cloud-folder-open",
         "path": "cloud-folder-open",
         "schema": {
             "type": "cloud-folder-open"
@@ -218,6 +227,7 @@
         "tabindex": 180,
         "messageKey": "cloud-folder-open"
     }, {
+        "id": "usb-open",
         "path": "usb-open",
         "schema": {
             "type": "usb-open"
@@ -226,6 +236,7 @@
         "tabindex": 190,
         "messageKey": "usb-open"
     }, {
+        "id": "launch-documorph",
         "path": "launch-documorph",
         "schema": {
             "type": "launch-documorph"
@@ -234,6 +245,7 @@
         "tabindex": 195,
         "messageKey": "launch-documorph"
     },  {
+        "id": "service-more",
         "path": "more",
         "schema": {
             "type": "more",
@@ -244,6 +256,7 @@
         "messageKey": "more",
         "tooltip": "<p>More options for changing the settings on this computer.</p>"
     }, {
+        "id": "service-save",
         "path": "save",
         "schema": {
             "type": "save",
@@ -254,6 +267,7 @@
         "messageKey": "save",
         "tooltip": "<p>Save your preferences so you can use them on other computers that have Morphic installed.</p>"
     }, {
+        "id": "service-undo",
         "path": "undo",
         "schema": {
             "type": "undo",
@@ -264,6 +278,7 @@
         "messageKey": "undo",
         "tooltip": "<p>Undo the last change you made in the Morphic QuickStrip.</p>"
     }, {
+        "id": "service-reset-all",
         "path": "resetAll",
         "schema": {
             "type": "resetAll",
@@ -274,6 +289,7 @@
         "messageKey": "resetAll",
         "tooltip": "<p>Resets computer to standard settings</p><p>(does not affect personal saved settings)</p>"
     }, {
+        "id": "service-close",
         "path": "close",
         "schema": {
             "type": "close",
